fix(lab): scope gem search to its market, and give the pool labels one home
Two changes; they overlap in ByVariant, which is why they share a commit.

1. Gem search was not scoped to the table it filled.

Searching a gem inside "Best Plays (20/20)" listed all four variants of it.
Each BestPlays table owned its own search, and a search fetched across every
variant then replaced that table's scoped rows wholesale — with the Var column
hidden in a variant-scoped table, the four results read as four identical rows
of the same gem. On the ALL tab there were four tables with four independent
boxes, so typing in one left the other three showing unrelated gems.

The search moves up to ByVariant, the component that knows the markets. One
query feeds every table through the same per-market and per-color filter the
ranked rows already go through. A market with no match now says so, which is
the useful answer: the gem does not sell at that level/quality.

This also fixes search in Dedication mode, which was silently dead — the fetch
never passed `mode`, so the server answered with normal-mode rows that carry no
`baseName` and could not match any pool tab.

BestPlays keeps only `searchActive`, the one thing a scoped table still needs
to know: a search names a gem explicitly, so its result shows even when that
gem is low-confidence and the toggle is off.

2. The Dedication pool labels had no single home.

"Skills" was written out at six call sites across both apps, so renaming it
meant finding all six. DEDICATION_POOLS and DEDICATION_POOL_LABELS now live in
each app's api.ts beside VARIANTS and DEDICATION_VARIANTS, and every surface
reads them from there. The canon value `skill` is unchanged — this is display
text only.

The label is now "Non-Transfigured" rather than "Skills". Not "Normal", which
was the first choice: the top bar already labels the lab mode Normal, and two
adjacent buttons both reading Normal would name two different things.
